Zabezpieczenia dostępu

+++ WebAPI/Controllers/V1/PostsController.cs
@@ -1,10 +1,12 @@
 ﻿using Application.Dto;
 using Application.Interfaces;
 using Domain.Entities;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.OData.Query;
 using Swashbuckle.AspNetCore.Annotations;
+using System.Security.Claims;
 using WebAPI.Filters;
 using WebAPI.Helpers;
 using WebAPI.Wrappers;
@@ -15,6 +17,7 @@ namespace WebAPI.Controllers.V1
     [Route("api/[controller]")]
     [ApiVersion("1.0")]
     [ApiController]
+    [Authorize]
     public class PostsController : ControllerBase
     {
         private readonly IPostService _postService;
@@ -32,6 +35,7 @@ namespace WebAPI.Controllers.V1
 
 
         [SwaggerOperation(Summary = "Retrives paged posts")]
+        [AllowAnonymous]
         [HttpGet]
         public async Task<IActionResult> Get([FromQuery] PaginationFilter paginationFilter, [FromQuery] SortingFilter sortingFilter, [FromQuery] string filterBy = "")
         {
@@ -48,6 +52,7 @@ namespace WebAPI.Controllers.V1
         [SwaggerOperation(Summary = "Retrives all posts")]
         [EnableQuery]
         [HttpGet("[action]")]
+        [AllowAnonymous]
         public IQueryable<PostDto> GetAll()
         {
             return _postService.GetAllPostsAsync();
@@ -67,6 +72,7 @@ namespace WebAPI.Controllers.V1
 
         [SwaggerOperation(Summary = "Retrives post by id")]
         [HttpGet("{id}")]
+        [AllowAnonymous]
         public async Task<IActionResult> Get(int id)
         {
             var post = await _postService.GetPostByIdAsync(id);
@@ -82,7 +88,7 @@ namespace WebAPI.Controllers.V1
         [HttpPost]
         public async Task<IActionResult> Create(CreatePostDto newPost)
         {
-            var post = await _postService.AddNewPostAsync(newPost);
+            var post = await _postService.AddNewPostAsync(newPost, User.FindFirstValue(ClaimTypes.NameIdentifier));
 
             return Created(uri: $"api/posts/{post.Id}", new Response<PostDto>(post));
         }
@@ -91,6 +97,15 @@ namespace WebAPI.Controllers.V1
         [HttpPut]
         public async Task<IActionResult> Update(UpdatePostDto updatePost)
         {
+            var userOwnsPost = await _postService.UserOwnsPostAsync(updatePost.Id, User.FindFirstValue(ClaimTypes.NameIdentifier));
+            if (!userOwnsPost)
+            {
+                return BadRequest(new Response<bool>
+                {
+                    Succeed = false,
+                    Message = "You do not own this post! You can't update the post."
+                });
+            }
             await _postService.UpdatePostAsync(updatePost);
 
             return NoContent();
@@ -98,9 +113,19 @@ namespace WebAPI.Controllers.V1
 
         [SwaggerOperation(Summary = "Delete a  post")]
         [HttpDelete("{id}")]
-        public IActionResult Delete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
-            _postService.DeletePostAsync(id);
+            var userOwnsPost = await _postService.UserOwnsPostAsync(id, User.FindFirstValue(ClaimTypes.NameIdentifier));
+            if (!userOwnsPost)
+            {
+                return BadRequest(new Response<bool>
+                {
+                    Succeed = false,
+                    Message = "You do not own this post! You can't delete the post."
+                });
+            }
+
+            await _postService.DeletePostAsync(id);
 
             return NoContent();
         }
